Extract modules from event persistence test to support reuse
